requirements.txt
openai-whisper
setuptools-rust
pymupdf
python-pptx
scikit-learn
